// ec/src/models/api/msm/ect/partNumber/SearchPartNumberResponse$search.ts
export interface PartNumberSpec {
	/**
	 * スペック項目コード
	 * - スペック項目のコード
	 * - example: E01
	 * - field groups: @search
	 */
	specCode: string;
	/**
	 * スペック項目名
	 * - スペック項目名
	 * - example: 長さ
	 * - field groups: @search
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @search
	 */
	specUnit?: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   1: テキストボタン形式
	 *   2: テキスト選択形式（1列）
	 *   3: テキスト選択形式（2列）
	 *   4: テキスト選択形式（3列）
	 *   5: 画像ボタン形式（1列）
	 *   6: 画像ボタン形式（2列）
	 *   7: 画像ボタン形式（3列）
	 *   8: 数値入力形式
	 *   9: リスト選択形式
	 *   10: ツリー(入れ子)選択形式
	 * - example: 1
	 * - field groups: @search
	 * - NOTE: 型番検索APIの仕様・寸法リストでは
	 *          11: 数値集約形式
	 *          12: 中央表示
	 *         は返却しない
	 *         型番検索APIの仕様・寸法リストでは
	 *         13～21の表示タイプは返却しない、該当するタイプは通常テンプレート向けに読み替える。（詳しい内容は別紙４参照）
	 */
	specViewType: SpecViewType;
	/**
	 * 開閉タイプ
	 * - スペック表示の開閉状態
	 *   1: 開
	 *   2: 閉
	 *   3: 開閉ボタン非表示
	 * - example: 1
	 * - field groups: @search
	 */
	openCloseType?: OpenCloseType;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @search
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType?: SupplementType;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @search
	 */
	specImageUrl?: string;
	/**
	 * 詳細HTML
	 * - 詳細ダイアログで表示するHTML
	 * - field groups: @search
	 */
	detailHtml?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @search
	 */
	specDescriptionImageUrl?: string;
	/**
	 * スペック注意喚起文
	 * - 注意喚起文の本文
	 * - field groups: @search
	 */
	specNoticeText?: string;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @search
	 */
	numericSpec?: NumericSpec;
	/**
	 * スペック値リスト
	 * - field groups: @search
	 * - NOTE: リスト内のすべてのスペック値が非表示の場合は、リスト(ボタン)を非表示
	 */
	specValueList: SpecValue[];
	/**
	 * 表示順
	 * - スペック項目の表示順
	 * - example: 1
	 * - field groups: @search
	 */
	displayOrder?: number;
}

/** 数値入力項目 */
export interface NumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @search
	 * - NOTE: そもそも数値入力欄が存在しない場合は、「数値入力項目」自体を返却しない
	 *         (空の値"{}"を返却)
	 */
	hiddenFlag: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @search
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲の選択範囲
	 * - field groups: @search
	 */
	specValueRangeList: SpecValueRange[];
}

/** 数値範囲 */
export interface SpecValueRange {
	/**
	 * 最小値
	 * - 指定可能な最小値
	 * - example: 1.0
	 * - field groups: @search
	 */
	minValue?: number;
	/**
	 * 最大値
	 * - 指定可能な最大値
	 * - example: 9.5
	 * - field groups: @search
	 */
	maxValue?: number;
	/**
	 * 刻み値
	 * - 値の刻み幅
	 * - example: 0.5
	 * - field groups: @search
	 */
	stepValue?: number;
}

/** スペック値 */
export interface SpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * スペック値画像URL
	 * - スペック値の画像URL
	 * - field groups: @search
	 * - NOTE: 設計書では必須
	 */
	specValueImageUrl?: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * 子スペック値リスト
	 * - 2段表示のスペック値リスト
	 * - field groups: @search
	 */
	childSpecValueList?: ChildSpecValue[];
	/**
	 * スペック値属性リスト
	 * - スペック値の属性情報
	 *   PU向けのスペック表示(表形式)に対応している場合に設定
	 * - field groups: @search
	 */
	specValueAttributeList?: SpecValueAttribute[];
}

/** 子スペック値 */
export interface ChildSpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
}

/** スペック値属性 */
export interface SpecValueAttribute {
	/**
	 * 属性リスト
	 * - 属性のリスト
	 * - field groups: @search
	 */
	attributeList?: Attribute[];
}
// ...
export interface AlterationSpec {
	/**
	 * スペック項目コード
	 * - スペック項目コード
	 * - example: E01
	 * - field groups: @search
	 */
	specCode: string;
	/**
	 * CAD用スペック項目コード
	 * - CAD用のスペック項目コード
	 * - example: LKC
	 * - field groups: @search
	 * - NOTE: SINUSでのCADデータ生成時に使用
	 */
	cadSpecCode?: string;
	/**
	 * スペック項目名
	 * - スペック項目名
	 * - example: 材質
	 * - field groups: @search
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @search
	 */
	specUnit?: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   2: テキスト選択形式（1列）
	 *   3: テキスト選択形式（2列）
	 *   4: テキスト選択形式（3列）
	 *   8: 数値入力形式
	 *   9: リスト選択形式
	 * - example: 8
	 * - field groups: @search
	 * - NOTE: 型番検索APIの追加工スペックリストでは
	 *           1: テキストボタン形式
	 *           5: 画像ボタン形式（1列）
	 *           6: 画像ボタン形式（2列）
	 *           7: 画像ボタン形式（3列）
	 *          10: ツリー(入れ子)選択形式
	 *          11: 数値集約形式
	 *          12: 中央表示
	 *         は返却しない
	 */
	specViewType: AlterationSpecViewType;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @search
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType: SupplementType;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @search
	 */
	specImageUrl?: string;
	/**
	 * 詳細HTML
	 * - 詳細ダイアログで表示するHTML
	 * - field groups: @search
	 */
	detailHtml?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @search
	 */
	specDescriptionImageUrl?: string;
	/**
	 * スペック注意喚起文
	 * - 注意喚起文の本文
	 * - field groups: @search
	 */
	specNoticeText?: string;
	/**
	 * 非表示フラグ
	 * - 規格式結果で隠すべきときで、ユーザに選択されていなければ隠す
	 *   0: 表示
	 *   1: 非表示
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 範囲値が埋まっていたり選択値がひとつでも選択されていた場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @search
	 */
	numericSpec: AlterationNumericSpec;
	/**
	 * スペック値リスト
	 * - field groups: @search
	 */
	specValueList: AlterationSpecValue[];
}

/** 数値入力項目 */
export interface AlterationNumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @search
	 * - NOTE: そもそも数値入力欄が存在しない場合は、「数値入力項目」自体を返却しない
	 *         (空の値"{}"を返却)
	 */
	hiddenFlag?: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @search
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲の選択範囲
	 * - field groups: @search
	 */
	specValueRangeList: SpecValueRange[];
}

/** 数値範囲 */
export interface SpecValueRange {
	/**
	 * 最小値
	 * - 指定可能な最小値
	 * - example: 1.0
	 * - field groups: @search
	 */
	minValue?: number;
	/**
	 * 最大値
	 * - 指定可能な最大値
	 * - example: 9.5
	 * - field groups: @search
	 */
	maxValue?: number;
	/**
	 * 刻み値
	 * - 値の刻み幅
	 * - example: 0.5
	 * - field groups: @search
	 */
	stepValue?: number;
}

/** スペック値 */
export interface AlterationSpecValue {
	/**
	 * スペック値
	 * - スペック値
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * CAD用スペック値
	 * - CAD用のスペック値
	 * - example: checked
	 * - field groups: @search
	 * - NOTE: SINUSでのCADデータ生成時に使用
	 */
	cadSpecValue?: string;
	/**
	 * スペック値画像URL
	 * - スペック値の画像URL
	 * - field groups: @search
	 */
	specValueImageUrl?: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * デフォルトフラグ
	 * - 型番確定時にデフォルトで選択されている追加工かどうか
	 *   0: デフォルトで選択されていない
	 *   1: デフォルトで選択されている
	 * - field groups: @search
	 * - NOTE: 参考
	 *         http://jp.misumi-ec.com/vona2/detail/221000197114/?HissuCode=YMH1SS04-2-S-B-NA-NB-[50-400%2F50]-1L-SA-CC&CategorySpec=00000049567%3A%3A50&Keyword=SS04-2-S-B-N-N-50-1L-SCC
	 */
	defaultFlag: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みの選択肢の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * スペック値属性リスト
	 * - スペック値の属性情報
	 *   PU向けのスペック表示(表形式)に対応している場合に設定
	 * - field groups: @search
	 */
	specValueAttributeList?: SpecValueAttribute[];
	/**
	 * 子スペック値リスト
	 * 設計書にはない項目。常に空？
	 */
	childSpecValueList: [];
}
// ...
export const SpecViewType = {
	TEXT_BUTTON: '1',
	TEXT_SINGLE_LINE: '2',
	TEXT_DOUBLE_LINE: '3',
	TEXT_TRIPLE_LINE: '4',
	IMAGE_SINGLE_LINE: '5',
	IMAGE_DOUBLE_LINE: '6',
	IMAGE_TRIPLE_LINE: '7',
	NUMERIC: '8',
	LIST: '9',
	TREE: '10',
} as const;
export type SpecViewType = typeof SpecViewType[keyof typeof SpecViewType];

/**
 * スペック項目表示タイプ
 * - スペックの表示種別
 *   2: テキスト選択形式（1列）
 *   3: テキスト選択形式（2列）
 *   4: テキスト選択形式（3列）
 *   8: 数値入力形式
 *   9: リスト選択形式
 */
export type AlterationSpecViewType = Extract<
	SpecViewType,
	'2' | '3' | '4' | '8' | '9'
>;

/**
 * 関連情報
 *   1: SDS(MSDS)
 *   2: データシート
 *   3: ミスミ定期便申し込み
 *   4: サンプル品提供申し込み
 */
export const RelatedLinkType = {
	MSDS: '1',
	DATA_SHEET: '2',
	MISUMI_REGULAR_SERVICE: '3',
	REQUEST_SAMPLE: '4',
} as const;
export type RelatedLinkType =
	typeof RelatedLinkType[keyof typeof RelatedLinkType];

/**
 * 補足表示タイプ
 * - スペックの補足表示種別
 *   1: 通常
 *   2: 外形図
 *   3: 詳細
 *   4: 拡大
 *   5: イラスト
 */
export const SupplementType = {
	NORMAL: '1',
	DRAWING: '2',
	DETAIL: '3',
	ZOOM_IMAGE: '4',
	ILLUSTRATION: '5',
} as const;
export type SupplementType = typeof SupplementType[keyof typeof SupplementType];