// ec/src/models/api/msm/ect/series/SearchSeriesResponse$search.ts
export interface SeriesSpec {
	/**
	 * スペック項目コード
	 * - スペック項目のコード値
	 * - example: E01
	 * - field groups: @search
	 * - NOTE: インチ・メトリックのスペック項目コード: unitType
	 */
	specCode: string;
	/**
	 * スペック項目名
	 * - スペック項目の名前
	 * - example: 形状
	 * - field groups: @search
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @search
	 */
	specUnit?: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   1: テキストボタン形式
	 *   2: テキスト選択形式（1列）
	 *   3: テキスト選択形式（2列）
	 *   4: テキスト選択形式（3列）、
	 *   5: 画像ボタン形式（1列）
	 *   6: 画像ボタン形式（2列）
	 *   7: 画像ボタン形式（3列）
	 *   8: 数値入力形式
	 *   9: リスト選択形式
	 *   11: 数値集約形式
	 *   12: 中央表示
	 * - example: 1
	 * - field groups: @search
	 * - NOTE: シリーズ検索APIでは
	 *          10: ツリー(入れ子)選択形式は返却しない、
	 *         タイプは「3:テキスト選択形式 (2列)」として返却
	 *         シリーズ検索APIでは
	 *         13～21の表示タイプは返却しない、
	 *         該当するタイプは通常テンプレート向けに読み替える。
	 *         （詳しい内容は別紙５参照）
	 */
	specViewType: SpecViewType;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @search
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType: SupplementType;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @search
	 */
	specImageUrl?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @search
	 */
	specDescriptionImageUrl?: string;
	/**
	 * 詳細HTML
	 * - 詳細ダイアログで表示するHTML
	 * - field groups: @search
	 */
	detailHtml?: string;
	/**
	 * 開閉タイプ
	 * - スペック表示の開閉状態
	 *   1: 開く
	 *   2: 閉じる
	 *   3: 開閉ボタン非表示
	 * - example: 1
	 * - field groups: @search
	 */
	openCloseType: OpenCloseType;
	/**
	 * スペック注意喚起文
	 * - スペックの注意文言
	 * - example: 各メーカーの精度基準をご確認下さい
	 * - field groups: @search
	 */
	specNoticeText?: string;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @search
	 */
	numericSpec?: NumericSpec;
	/**
	 * スペック値リスト
	 * - スペック値のリスト
	 * - field groups: @search
	 * - NOTE: リスト内のすべてのスペック値が非表示の場合は、リスト(ボタン)を非表示
	 */
	specValueList: SpecValue[];
	/**
	 * 表示順
	 * - スペック項目の表示順
	 * - field groups: @search
	 * - example: 1
	 */
	displayOrder: number;
}

/** 数値入力項目 */
export interface NumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @search
	 * - NOTE: そもそも数値入力欄が存在しない場合は、「数値入力項目」自体を返却しない
	 */
	hiddenFlag: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @search
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲のリスト
	 * - field groups: @search
	 */
	specValueRangeList: SpecValueRange[];
}

/** 数値範囲 */
export interface SpecValueRange {
	/**
	 * 最小値
	 * - 指定可能な最小値
	 * - example: 1.0
	 * - field groups: @search
	 */
	minValue?: number;
	/**
	 * 最大値
	 * - 指定可能な最大値
	 * - example: 9.5
	 * - field groups: @search
	 */
	maxValue?: number;
	/**
	 * 刻み値
	 * - 値の刻み幅
	 * - example: 0.5
	 * - field groups: @search
	 */
	stepValue?: number;
}

/** スペック値 */
export interface SpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 100
	 * - field groups: @search
	 * - NOTE: インチ・メトリックのスペック値コード:
	 *         　1: metric
	 *           2: inch
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 加工なし
	 * - field groups: @search
	 * - NOTE: インチ・メトリックのスペック値表示文言:
	 *         　1: metric
	 *           2: inch
	 */
	specValueDisp: string;
	/**
	 * スペック値画像URL
	 * - スペック値の画像URL
	 * - example: //xxx/PHOTO/10302634310.jpg
	 * - field groups: @search
	 */
	specValueImageUrl?: string;
	/**
	 * 非表示フラグ
	 * - このスペック値を表示するかどうか
	 *   0: 表示
	 *   1: 非表示
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * 子スペック値リスト
	 * - 2段表示のスペック値リスト
	 * - field groups: @search
	 * - NOTE: スペック値タイプが2,3のときのみ、値がセットされる
	 */
	childSpecValueList: ChildSpecValue[];
}

export interface ChildSpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
}
// ...
export const SpecViewType = {
	TEXT_BUTTON: '1',
	TEXT_SINGLE_LINE: '2',
	TEXT_DOUBLE_LINE: '3',
	TEXT_TRIPLE_LINE: '4',
	IMAGE_SINGLE_LINE: '5',
	IMAGE_DOUBLE_LINE: '6',
	IMAGE_TRIPLE_LINE: '7',
	NUMERIC: '8',
	LIST: '9',
	TREE: '10',
	AGGREGATION: '11',
	CENTER: '12',
} as const;
export type SpecViewType = typeof SpecViewType[keyof typeof SpecViewType];

/**
 * 補足表示タイプ
 * - スペックの補足表示種別
 *   1: 通常
 *   2: 外形図
 *   3: 詳細
 *   4: 拡大
 *   5: イラスト
 */
export const SupplementType = {
	NORMAL: '1',
	DRAWING: '2',
	DETAIL: '3',
	ZOOM_IMAGE: '4',
	ILLUSTRATION: '5',
} as const;
export type SupplementType = typeof SupplementType[keyof typeof SupplementType];