// ec/src/models/api/msm/ect/categoryPartNumber/SearchCategoryPartNumberResponse.ts
export interface PartNumberSpec {
	/**
	 * スペック項目コード
	 * - スペック項目のコード
	 * - example: E01
	 * - field groups: @default, @search, @detail
	 */
	specCode: string;
	/**
	 * 追加工フラグ
	 * - C項目追加工の場合に設定
	 *   0: C項目追加工以外のスペック
	 *   1: C項目追加工
	 * - example: 1
	 * - field groups: @default, @search, @detail
	 */
	alterationFlag?: Flag;
	/**
	 * スペック項目名
	 * - スペック項目名
	 * - example: 長さ
	 * - field groups: @default, @search, @detail
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @default, @search, @detail
	 */
	specUnit: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   8:数値入力形式
	 *   9:リスト選択形式
	 *   11:数値集約形式
	 *   13.テキスト選択形式（3列）
	 *   14.テキスト選択形式（6列）
	 *   15.テキスト選択形式（9列）
	 *   16.イラスト選択形式（2列）
	 *   17.イラスト選択形式（4列）
	 *   18.イラスト選択形式（6列）
	 *   19:ツリー(入れ子)選択形式（3列）
	 *   20:ツリー(入れ子)選択形式（6列）
	 *   21:プルダウン選択形式
	 * - example: 1
	 * - field groups: @default, @search, @detail
	 */
	specViewType: string;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @default, @search, @detail
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType: string;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @default, @search, @detail
	 */
	specImageUrl?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @default, @search, @detail
	 */
	specDescriptionImageUrl?: string;
	/**
	 * スペック注意喚起文
	 * - 注意喚起文の本文
	 * - field groups: @default, @search, @detail
	 */
	specNoticeText: string;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @default, @search, @detail
	 */
	numericSpec: NumericSpec;
	/**
	 * スペック値リスト
	 * - field groups: @default, @search, @detail
	 */
	specValueList: SpecValue[];
	/**
	 * 表示順
	 * - スペック項目の表示順
	 * - example: 1
	 * - field groups: @default, @search, @detail
	 */
	displayOrder: number;
}

/** 数値入力項目 */
export interface NumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @default, @search, @detail
	 */
	hiddenFlag: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @default, @search, @detail
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲の選択範囲
	 * - field groups: @default, @search, @detail
	 */
	specValueRangeList: SpecValueRange[];
}

/** 数値範囲 */
export interface SpecValueRange {
	/**
	 * 最小値
	 * - 指定可能な最小値
	 * - example: 1.0
	 * - field groups: @default, @search, @detail
	 */
	minValue?: number;
	/**
	 * 最大値
	 * - 指定可能な最大値
	 * - example: 9.5
	 * - field groups: @default, @search, @detail
	 */
	maxValue?: number;
	/**
	 * 刻み値
	 * - 値の刻み幅
	 * - example: 0.5
	 * - field groups: @default, @search, @detail
	 */
	stepValue?: number;
}

/** スペック値 */
export interface SpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @default, @search, @detail
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @default, @search, @detail
	 */
	specValueDisp: string;
	/**
	 * スペック値画像URL
	 * - スペック値の画像URL
	 * - field groups: @default, @search, @detail
	 */
	specValueImageUrl: string;
	/**
	 * 非表示フラグ
	 * - field groups: @default, @search, @detail
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @default, @search, @detail
	 */
	selectedFlag: Flag;
	/**
	 * 子スペック値リスト
	 * - 2段表示のスペック値リスト
	 * - field groups: @default, @search, @detail
	 */
	childSpecValueList: ChildSpecValue[];
}

export interface ChildSpecValue {
	specValue: string;
	hiddenFlag: Flag;
	selectedFlag: Flag;
	specValueDisp: string;
}